address PR feedback and remove unused imports
